requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
selenium==4.15.2
2captcha-python==1.1.3
pydantic==2.5.0
python-dotenv==1.0.0
